backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-cors-headers>=4.6.0",
    "django>=5.1.4",
    "beautifulsoup4>=4.12.3",
    "googlemaps>=4.10.0",
    "utils>=1.0.2",
    "python-decouple>=3.8",
]
